sukesh/backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
PyGithub
pdfkit
reportlab
weasyprint
authlib
pytest